+++ inst/rstudio/templates/project/proj_fls/0_retrieve-data.R
@@ -1,0 +1,2 @@
+source("shared.R")
+reschola::gd_download_folder(gd_url, overwrite = F, files_from_subfolders = T)